Components for Choosing quizes, Home page and Rules updated/created
